Add files via upload
The Library Management System is a web-based application designed to streamline the process of managing a library's resources and operations. It provides an intuitive and user-friendly interface for librarians to handle tasks such as adding books, managing borrowers,  and generating reports in cart. The system ensures efficient library operations and enhances the overall user experience.